+++ core/src/main/resources/phoenix/create_tables.sql
@@ -9,7 +9,8 @@ CREATE TABLE IF NOT EXISTS topology_editor_metadata ("topologyId" BIGINT NOT NUL
 CREATE TABLE IF NOT EXISTS tag ("id" BIGINT NOT NULL, "name" VARCHAR(256), "description" VARCHAR(256), "timestamp" BIGINT, CONSTRAINT pk PRIMARY KEY ("id"))
 CREATE TABLE IF NOT EXISTS tag_storable_mapping ("tagId" BIGINT NOT NULL, "storableNamespace" VARCHAR(32) NOT NULL, "storableId" BIGINT NOT NULL, CONSTRAINT pk PRIMARY KEY ("tagId", "storableNamespace", "storableId"))
 CREATE TABLE IF NOT EXISTS streaminfo ("id" BIGINT NOT NULL, "streamId" VARCHAR(256), "fieldsData" VARCHAR, "timestamp" BIGINT, CONSTRAINT pk PRIMARY KEY ("id"))
-CREATE TABLE IF NOT EXISTS sequence_table ("id" VARCHAR, "datasources" BIGINT, "datafeeds" BIGINT, "parser_info" BIGINT, "files" BIGINT, "topologies" BIGINT, "topology_components" BIGINT, "tag" BIGINT,  "streaminfo" BIGINT, CONSTRAINT pk PRIMARY KEY ("id"))
+CREATE TABLE IF NOT EXISTS notifierinfo ("id" BIGINT NOT NULL, "name" VARCHAR(256), "jarFileName" VARCHAR(256), "className" VARCHAR(256), "propertiesData" VARCHAR, "fieldValuesData" VARCHAR, "timestamp" BIGINT, CONSTRAINT pk PRIMARY KEY ("id"))
+CREATE TABLE IF NOT EXISTS sequence_table ("id" VARCHAR, "datasources" BIGINT, "datafeeds" BIGINT, "parser_info" BIGINT, "files" BIGINT, "topologies" BIGINT, "topology_components" BIGINT, "tag" BIGINT,  "streaminfo" BIGINT, "notifierinfo" BIGINT, CONSTRAINT pk PRIMARY KEY ("id"))
 CREATE SEQUENCE IF NOT EXISTS datasources_sequence
 CREATE SEQUENCE IF NOT EXISTS datafeeds_sequence
 CREATE SEQUENCE IF NOT EXISTS parser_info_sequence
@@ -18,3 +19,4 @@ CREATE SEQUENCE IF NOT EXISTS topology_components_sequence
 CREATE SEQUENCE IF NOT EXISTS tag_sequence
 CREATE SEQUENCE IF NOT EXISTS files_sequence
 CREATE SEQUENCE IF NOT EXISTS streaminfo_sequence
+CREATE SEQUENCE IF NOT EXISTS notifierinfo_sequence